web: Refactor sample selection
